Create Motor and spec interfaces in device_interfaces.py
